requirements.txt
logzero
python-docx
pytest
beautifulsoup4
pandas
openpyxl
spacy
jinja2
pypdf
pdfminer
pdftitle
"pdfminer.six"
pymupdf
xhtml2pdf

